+++ NEWS.md
@@ -1,3 +1,7 @@
+# ggrefine (development)
+
+* Change panel grey in `theme_grey` to "grey95".
+
 # ggrefine 0.2.0
 
 * Renamed theme functions, and removed `theme_oat`.